Using map and removing Any type attrs

+++ app/src/main/java/com/rafaellsdev/cryptocurrencyprices/feature/home/repository/model/CurrencyResponse.kt
@@ -10,7 +10,7 @@ data class CurrencyResponse(
     @SerializedName("current_price")
     val currentPrice: Double?,
     @SerializedName("price_change_percentage_24h")
-    val priceChangePercentage: Float?,
+    val priceChangePercentage: Double?,
     @SerializedName("high_24h")
     val highPrice: Double?,
     @SerializedName("low_24h")
